requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
discord.py==1.4.1
lxml==4.5.2
pycparser==2.20
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyOpenSSL==19.1.0
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python3-xlib==0.15
requests==2.22.0
